+++ chatbot-listen.js
@@ -1,35 +1,75 @@
 var speak = require("speakeasy-nlp");
 var levenshtein = require('fast-levenshtein');
+var ChatContext = require('./lib/chat-context.js');
 var _ = require('underscore');
+var regexps = require('./lib/helpers/regexps.js');
 var debug = false;
 
 module.exports = function(RED) {
 
+  var fixedWords = ['yes', 'no', 'on', 'off'];
+  var tokenVariables = ['{{email}}'];
+
+  function isFixedWord(word) {
+    return _.contains(fixedWords, word);
+  }
+
+  function isTokenVariable(word) {
+    return _.contains(tokenVariables, word);
+  }
+
+  /**
+   * @method getDistance
+   * Get the Levenshtein distance based on the length of the word, for length = 2 distance must be 0 or "off" and "on"
+   * will be confused
+   */
   function getDistance(word) {
     if (word.length <= 2) {
       return 0
-    } else if (word.length <=4) {
+    } else if (word.length <= 4) {
       return 1;
     } else {
       return 2;
     }
   }
 
-  function matchSentence(sentence, words) {
-
+  /**
+   * @method matchSentence
+   * Given the analysis to a sentence, check the all the provided words match with the sentence
+   * @param {Object} sentence
+   * @param {Array} words
+   * @param {ChatContext} chatContext
+   */
+  function matchSentence(sentence, words, chatContext) {
+    // convert all words lowercase
     words = _(words).map(function(word) {
       return word.toLowerCase();
     });
-    var exactWords = ['yes', 'no', 'on', 'off'];
+
 
     debug && console.log('tokens - ', sentence.tokens);
+    debug && console.log('analysis - ', sentence);
 
     // scan the words, all must be present
     var result = _(words).all(function(word) {
 
-      if (_.contains(exactWords, word)) {
+      if (isFixedWord(word)) {
         debug && console.log('contain exact', _.contains(sentence.tokens, word));
         return _.contains(sentence.tokens, word);
+      } else if (isTokenVariable(word)) {
+        // get the right matcher
+        var variable = word.replace('{{', '').replace('}}', '');
+        var test = regexps[variable];
+        var found = null;
+        // search if in the tokens something is matching
+        var matched = _(sentence.tokens).some(function(token) {
+          return (found = test(token)) != null;
+        });
+        // store in chat context if any
+        if (found != null) {
+          chatContext.set(variable, found);
+        }
+        return matched;
       } else {
         return _(sentence.tokens).some(function(token) {
           debug && console.log('* Levenshtein ', token, word, levenshtein.get(token, word) <= getDistance(word));
@@ -51,6 +91,10 @@ module.exports = function(RED) {
 
     this.on('input', function(msg) {
       var sentences = node.sentences;
+      var originalMessage = msg.originalMessage;
+      var chatId = msg.payload.chatId || (originalMessage && originalMessage.chat.id);
+      var context = node.context();
+      var chatContext = context.flow.get('chat:' + chatId) || ChatContext(chatId);
 
       // exit if not string
       if (!_.isString(msg.payload.content)) {
@@ -60,30 +104,15 @@ module.exports = function(RED) {
 
       // see if one of the rules matches
       var matched = _(sentences).any(function(sentence) {
-        if (sentence.type == 'str') {
-          if (msg.payload.content == sentence.value) {
-            return true;
-          }
-        } else {
-          // check if valid json
-          var words = null;
 
-          try {
-            words = JSON.parse(sentence.value);
-          } catch(err) {
-            node.error('Error parsing list of words. Only valid JSON like ["word1","word2"], remember to use " and not \'.');
-          }
-          // check if array
-          if (!_.isArray(words)) {
-            node.error('List of words must be an array. Something like ["word1","word2"], remember to use " and not \'.');
-            return;
-          }
-          debug && console.log('---- START analisys ' + msg.payload.content);
-          // analize sentence
-          var analysis = speak.classify(msg.payload.content);
-          // send message if the sentence is matched, otherwise stop here
-          return matchSentence(analysis, words);
-        }
+        // check if valid json
+        var words = sentence.split(',');
+
+        debug && console.log('---- START analysis ' + msg.payload.content);
+        // analize sentence
+        var analysis = speak.classify(msg.payload.content);
+        // send message if the sentence is matched, otherwise stop here
+        return matchSentence(analysis, words, chatContext);
       });
       // pass through the message if any of the above matched
       if (matched) {