+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Node-Red-Contrib-ChatBot Changelog
-## 0.3.14
+## 0.3.16
+* Improved UI for listen node
+* Some documentation for listen node
+* Added speacial token {{email}} to listen node
+* Improved getting started section of docs
+
+## 0.3.15
 * Receive image in Telegram
 * Dates in Telegram are now momentjs objects
 * Parse node now extracts photo from message payload